Create add to cart method
Create a add to cart method in the cartController file then associate the file to the cartService file and implement it in the serviceImpl,
Create new cart, find the product by product Id, then create a custom repository method to find cart item by cartId and productId,
Add check points for the cart then create new cart items and set the products, carts etc.,
Create AuthUtil file in the util folder to authenticate the user who's logged in, logged out, signed in.
Create a function to proceed add to cart functionality to the signed in user,
Add a query annotation to specifty the custom repository to find cart item by cartId and productId.