Save the frontend cart items in the backend
Create controller to save the cart items using post method to the DB,
Change the cart item dto to product id, quantity which will be the fields that interacts with user in the frontend,
Create a delete query that delete cart items by cart id in the cart item repository,
Add the bean annotation in the web config file to use it globally,
Implemented transactional cart creation/update with item persistence and price calculation : (in the cart service implementation file)
This method creates or updates a cart for the logged-in user. If no cart exists, a new one is created;
otherwise, existing cart items are removed.
It then iterates through the provided cart items, fetches product details, calculates total price, and saves each item in the cart.
Finally, it updates the cart’s total price and saves the cart. The method runs within a transactional scope to ensure consistency.